% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hash.R
\docType{package}
\name{hashr-package}
\alias{hashr-package}
\title{Hash R Objects Quickly}
\description{
This package exports Paul Hsies's \code{SuperFastHash} C-code to R.
It can be used to hash either whole R objects or, for vectors or lists,
R objects can be hashed recursively so one obtains a set of hash values
that is stored in a structure equivalent to the input.
}

